Move Payloads module and source tests to a directory
