+++ tests/nosetests/pyanaconda_tests/modules/payloads/payload/module_payload_dnf_manager_test.py
