+++ tests/nosetests/pyanaconda_tests/modules/payloads/payload/module_payload_shared.py
