+++ tests/nosetests/pyanaconda_tests/modules/payloads/source/source_base_test.py
