+++ tests/nosetests/pyanaconda_tests/modules/payloads/source/source_cdn_test.py
