+++ tests/nosetests/pyanaconda_tests/modules/payloads/source/source_cdrom_test.py
