+++ tests/nosetests/pyanaconda_tests/modules/payloads/source/source_closest_mirror_test.py
