+++ tests/nosetests/pyanaconda_tests/modules/payloads/source/source_factory_test.py
