+++ tests/nosetests/pyanaconda_tests/modules/payloads/source/source_harddrive_test.py
