+++ tests/nosetests/pyanaconda_tests/modules/payloads/source/source_hmc_test.py
