+++ tests/nosetests/pyanaconda_tests/modules/payloads/source/source_live_image_test.py
