+++ tests/nosetests/pyanaconda_tests/modules/payloads/source/source_live_os_test.py
