+++ tests/nosetests/pyanaconda_tests/modules/payloads/source/source_repo_files_test.py
