+++ tests/nosetests/pyanaconda_tests/modules/payloads/source/source_rpm_ostree_test.py
