+++ tests/nosetests/pyanaconda_tests/modules/payloads/source/source_utils_test.py
